refactoring

+++ MinecraftServerEngine/AbstractPlayer.cs
@@ -220,17 +220,17 @@ namespace MinecraftServerEngine
 
         internal override void _EmitParticles(
             Particle particle, Vector v,
-            float speed, int count,
-            float r, float g, float b)
+            double speed, int count,
+            double r, double g, double b)
         {
-            System.Diagnostics.Debug.Assert(r >= 0.0D);
-            System.Diagnostics.Debug.Assert(r <= 1.0D);
-            System.Diagnostics.Debug.Assert(g >= 0.0D);
-            System.Diagnostics.Debug.Assert(g <= 1.0D);
-            System.Diagnostics.Debug.Assert(b >= 0.0D);
-            System.Diagnostics.Debug.Assert(b <= 1.0D);
-            System.Diagnostics.Debug.Assert(speed >= 0.0F);
-            System.Diagnostics.Debug.Assert(speed <= 1.0F);
+            System.Diagnostics.Debug.Assert(r >= 0.0);
+            System.Diagnostics.Debug.Assert(r <= 1.0);
+            System.Diagnostics.Debug.Assert(g >= 0.0);
+            System.Diagnostics.Debug.Assert(g <= 1.0);
+            System.Diagnostics.Debug.Assert(b >= 0.0);
+            System.Diagnostics.Debug.Assert(b <= 1.0);
+            System.Diagnostics.Debug.Assert(speed >= 0.0);
+            System.Diagnostics.Debug.Assert(speed <= 1.0);
             System.Diagnostics.Debug.Assert(count >= 0);
 
             System.Diagnostics.Debug.Assert(_disposed == false);
@@ -242,7 +242,7 @@ namespace MinecraftServerEngine
 
             if (Connected == true)
             {
-                Conn.EmitParticles(particle, v, speed, count, r, g, b);
+                Conn.EmitParticles(particle, v, (float)speed, count, (float)r, (float)g, (float)b);
             }
         }
